Refactor code: eluminate code duplicating, fix arguments
