+++ project/cfpq_algorithms.py
@@ -1,15 +1,18 @@
 from typing import Set
+from collections.abc import Callable
 
 from project.graphs_lib import LABEL
 from project.cfg import cfg_to_wcnf
 
 from networkx import Graph
 from pyformlang.cfg import CFG, Variable
+from scipy.sparse import lil_matrix
 
 
-def helling_request(
+def cfpq_request(
     graph: Graph,
     request: CFG,
+    algorithm: Callable[Graph, CFG, Set],
     start_vertices: Set = None,
     final_vertices: Set = None,
     start_variable: Variable = Variable("S"),
@@ -41,7 +44,7 @@ def helling_request(
     if final_vertices is None:
         final_vertices = graph.nodes
 
-    transitive_closure = constrained_transitive_closure(graph, request)
+    transitive_closure = algorithm(graph, request)
 
     res = {
         (start_node, final_node)
@@ -52,6 +55,25 @@ def helling_request(
     }
 
     return res
+
+
+def _prepare_wcfg_for_algorithm(wcnf: CFG) -> (set, set, set):
+
+    epsilon_prods = set()
+    term_prods = set()
+    var_prods = set()
+
+    for prod in wcnf.productions:
+        prod_body_len = len(prod.body)
+
+        if prod_body_len == 0:
+            epsilon_prods.add(prod.head)
+        elif prod_body_len == 1:
+            term_prods.add(prod)
+        else:  # prod_body_len == 2
+            var_prods.add(prod)
+
+    return epsilon_prods, term_prods, var_prods
 
 
 def constrained_transitive_closure(graph: Graph, cfg: CFG) -> Set:
@@ -71,20 +93,7 @@ def constrained_transitive_closure(graph: Graph, cfg: CFG) -> Set:
     """
 
     wcnf = cfg_to_wcnf(cfg)
-
-    epsilon_prods = set()
-    term_prods = set()
-    var_prods = set()
-
-    for prod in wcnf.productions:
-        prod_body_len = len(prod.body)
-
-        if prod_body_len == 0:
-            epsilon_prods.add(prod.head)
-        elif prod_body_len == 1:
-            term_prods.add(prod)
-        else:  # prod_body_len == 2
-            var_prods.add(prod)
+    epsilon_prods, term_prods, var_prods = _prepare_wcfg_for_algorithm(wcnf)
 
     res = {(node, var.value, node) for node in graph.nodes for var in epsilon_prods} | {
         (first_node, prod.head.value, second_node)
@@ -125,3 +134,37 @@ def constrained_transitive_closure(graph: Graph, cfg: CFG) -> Set:
         res |= tmp
 
     return res
+
+
+def matrix_closure(graph: Graph, cfg: CFG) -> Set:
+
+    wcnf = cfg_to_wcnf(cfg)
+    epsilon_prods, term_prods, var_prods = _prepare_wcfg_for_algorithm(wcnf)
+
+    nodes = list(graph.nodes)
+    indexes_nodes = {node: i for i, node in enumerate(nodes)}
+    n = graph.number_of_nodes()
+
+    matrices = {var: lil_matrix((n, n), dtype=bool) for var in wcnf.variables}
+
+    for i in range(n):
+        for var in epsilon_prods:
+            matrices[var][i, i] = True
+
+    for u, v, label in graph.edges(data="label"):
+        i, j = indexes_nodes[u], indexes_nodes[v]
+        for var in term_prods:
+            matrices[var.head][i, j] |= label == var.body[0].value
+
+    while True:
+        old_nnz = sum([v.nnz for v in matrices.values()])
+        for prod in var_prods:
+            matrices[prod.head] += matrices[prod.body[0]] @ matrices[prod.body[1]]
+        if old_nnz == sum([value.nnz for value in matrices.values()]):
+            break
+
+    return set(
+        (nodes[i], v, nodes[j])
+        for v, mat in matrices.items()
+        for i, j in zip(*mat.nonzero())
+    )
